+++ HISTORY.rst
@@ -2,6 +2,11 @@
 History
 =======
 
+Unreleased
+----------
+
+* Fix bug: opening gzipped files with Py3 now works
+
 0.2.0 (2017-11-25)
 ------------------